requirements.txt
google-adk
playwright
python-dotenv
pydantic
litellm